execution/protocol, execution/protocol/rules: remove BlockPostValidation shim, move its tests to rules
